Refactor TraceTask to split up lighting calculation and allow easier recursion; add initial recursion and reflectance support; add simple stochastic antialiasing (looks awful :( )

+++ src/main/java/org/timothyb89/trace/math/Camera.java
@@ -16,22 +16,22 @@ public class Camera {
 	private final Vector viewUp;
 	private final double focalLength;
 	private final int[] bounds; // -u, -v, +u, +v
-	
+
 	/**
 	 * The View Plane Normal axis (VPN), looking directly down the camera.
 	 */
 	private Vector n;
-	
+
 	/**
 	 * The camera "x" axis.
 	 */
 	private Vector u;
-	
+
 	/**
 	 * The camera "y" axis.
 	 */
 	private Vector v;
-	
+
 	/**
 	 * The base vector for {@code loc()}, FP + dN.
 	 */
@@ -44,10 +44,10 @@ public class Camera {
 		this.focalPoint = focalPoint;
 		this.lookAtPoint = lookAtPoint;
 		this.viewUp = viewUp;
-		
+
 		this.focalLength = focalLength;
 		this.bounds = bounds;
-		
+
 		resolveAxes();
 	}
 
@@ -58,16 +58,16 @@ public class Camera {
 	public int height() {
 		return bounds[3] - bounds[1] + 1; // (v) - (-v), inclusive range
 	}
-	
+
 	private void resolveAxes() {
 		n = lookAtPoint.copy().sub(focalPoint).normalize();
 		u = n.copy().cross(viewUp).normalize();
 		v = u.copy().cross(n);
-		
+
 		baseLoc = focalPoint.copy().add(n.copy().scale(focalLength));
 	}
-	
-	public Vector lensPoint(int row, int col) {
+
+	public Vector lensPoint(double row, double col) {
 		return baseLoc.copy()              // (F + dN)
 				.add(u.copy().scale(row))  // + rU
 				.add(v.copy().scale(col)); // + cV
@@ -76,5 +76,5 @@ public class Camera {
 	public int[] bounds() {
 		return bounds;
 	}
-	
+
 }
